requirements.txt
flask==2.3.3
torch==2.1.2
torchvision==0.16.2
pillow==10.1.0
